refactor: move syncset and syncmap to internal path
